Be stricter with TLS configuration for ntpd constraints.
We already require TLSv1.2 so it does not make sense to be liberal with the
cipher suites that we allow. Additionally, it is potentially dangerous to
disable certificate verification when no CA data is available (which is
currently an impossible case to reach).

Also ensure we check the return value from tls_config_set_ca_mem() (as
spotted by tb@).

ok kn@ tb@